requirements.txt
ahk
ahk-binary
anthropic
dacite
dashscope
dataclass_wizard
dxcam; sys_platform == "win32"
google-genai
gymnasium==1.1.1
imageio[ffmpeg]
langchain-chroma
langchain-openai
mcp[cli]
mss
Multi-Template-Matching==1.6.6
numpy
omegaconf
openai
opencv-contrib-python
pandas
pillow>=10.3.0
pyautogui
pydantic
pydirectinput; sys_platform == "win32"
pygame
pyobjc-framework-Cocoa; sys_platform == "darwin"
pyobjc-framework-Quartz; sys_platform == "darwin"
pywin32; sys_platform == "win32"
pyyaml
rich
screeninfo
tenacity
termcolor
tiktoken
tqdm
transformers
volcengine-python-sdk